+++ CHANGELOG.md
@@ -5,11 +5,15 @@ What's going on?
 ## Future ideas
 
 * add a nice progress bar to give a better feedback to the user
-* add a `--stats` flag to show some conversion data after the entire process
 * add a `--filter` flag to filter images that must be resized, by file name
 * create executables for different platforms (Linux, Windows, OSX)
-* separate code and responsibilities in a better way
 * write some tests
+
+## 0.3
+
+* added a `--verbose` flag that shows more info about the resize process
+* added an error message if no images are found
+* resized images are now saved in the "resized" folder as a default behavior
 
 ## 0.2